+++ packages/driver/cypress/plugins/server.js
@@ -41,13 +41,17 @@ const createApp = (port) => {
     .send('<html><body>hello there</body></html>')
   })
 
-  app.get('/redirect', (req, res) => {
+  app.get('/status-code', (req, res) => {
+    res.sendStatus(req.query.code || 200)
+  })
+
+  app.all('/redirect', (req, res) => {
     if (req.query.chunked) {
       res.setHeader('transfer-encoding', 'chunked')
       res.removeHeader('content-length')
     }
 
-    res.statusCode = 301
+    res.statusCode = Number(req.query.code || 301)
     res.setHeader('Location', req.query.href)
     res.end()
   })