+++ packages/net-stubbing/lib/server/intercept-response.ts
@@ -1,7 +1,7 @@
 import _ from 'lodash'
-import concatStream from 'concat-stream'
+import { concatStream, httpUtils } from '@packages/network'
 import Debug from 'debug'
-import { Readable } from 'stream'
+import { Readable, PassThrough } from 'stream'
 
 import {
   ResponseMiddleware,
@@ -65,17 +65,30 @@ export const InterceptResponse: ResponseMiddleware = function () {
 
   this.makeResStreamPlainText()
 
-  this.incomingResStream.pipe(concatStream((resBody) => {
-    res.body = resBody.toString()
+  new Promise((resolve) => {
+    if (httpUtils.responseMustHaveEmptyBody(this.req, this.incomingRes)) {
+      resolve('')
+    } else {
+      this.incomingResStream.pipe(concatStream((resBody) => {
+        resolve(resBody)
+      }))
+    }
+  })
+  .then((body) => {
+    const pt = this.incomingResStream = new PassThrough()
+
+    pt.end(body)
+
+    res.body = String(body)
     emitReceived()
-  }))
 
-  if (!backendRequest.waitForResponseContinue) {
-    this.next()
-  }
+    if (!backendRequest.waitForResponseContinue) {
+      this.next()
+    }
 
-  // this may get set back to `true` by another route
-  backendRequest.waitForResponseContinue = false
+    // this may get set back to `true` by another route
+    backendRequest.waitForResponseContinue = false
+  })
 }
 
 export async function onResponseContinue (state: NetStubbingState, frame: NetEventFrames.HttpResponseContinue) {